+++ R/zzz.R
@@ -17,3 +17,19 @@ NULL
 #'
 #' @keywords internal
 NULL
+
+#' Shared parameters for get functions
+#'
+#' @name shared-get
+#'
+#' @param study An OmicAnalyzer study. Either an object of class \code{oaStudy}
+#'   or the name of an installed study package.
+#' @param modelID Filter by modelID
+#' @param testID Filter by testID
+#' @param annotationID Filter by annotation
+#' @param termID Filter by termID
+#' @param featureID Filter by featureID
+#' @param plotID Filter by plotID
+#'
+#' @keywords internal
+NULL